+++ misc/metrics/examples/metrics/main.rs
@@ -53,7 +53,7 @@ use futures::executor::block_on;
 use futures::stream::StreamExt;
 use libp2p::core::Multiaddr;
 use libp2p::metrics::{Metrics, Recorder};
-use libp2p::swarm::{NetworkBehaviour, SwarmEvent};
+use libp2p::swarm::SwarmEvent;
 use libp2p::{identity, ping, NetworkBehaviour, PeerId, Swarm};
 use libp2p_swarm::keep_alive;
 use log::info;