+++ protocols/autonat/CHANGELOG.md
@@ -1,6 +1,10 @@
 # 0.8.0 [unreleased]
+ 
+- Update to `libp2p-core` `v0.37.0`.
 
 - Update to `libp2p-swarm` `v0.40.0`.
+
+- Update to `libp2p-request-response` `v0.22.0`.
 
 # 0.7.0