+++ protocols/rendezvous/CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.10.0 [unreleased]
 
+- Update to `libp2p-core` `v0.37.0`.
+
 - Update to `libp2p-swarm` `v0.40.0`.
 
 # 0.9.0